+++ lab/reports/__init__.py
@@ -70,38 +70,55 @@ def function_name(f):
     return names.get(f.__name__, f.__name__)
 
 
+def get_aggregation_function(function, functions):
+    """
+    Code for backwards compatibility.
+    """
+    if function and functions:
+        logging.critical(
+            'You cannot use "function" and "functions" kwargs for '
+            'Attribute at the same time.')
+    elif functions:
+        tools.show_deprecation_warning(
+            '"functions" kwarg for Attribute is deprecated. Use '
+            '"function" instead.')
+        if len(functions) > 1:
+            logging.critical(
+                'Using multiple aggregation functions is unsupported.')
+        return functions[0]
+    else:
+        return function
+
+
 class Attribute(str):
     """A string subclass for attributes in reports."""
     def __new__(cls, name, **kwargs):
         return str.__new__(cls, name)
 
     def __init__(
-            self, name, absolute=False, min_wins=True, functions=sum,
-            scale=None, digits=2):
+            self, name, absolute=False, min_wins=True, function=None,
+            functions=None, scale=None, digits=2):
         """
-        Use this class if your **own** attribute needs a non-default
+        Use this class if your **custom** attribute needs a non-default
         value for:
 
-        * *absolute*: If False, only include tasks for which all task
+        * *absolute*: if False, only include tasks for which all task
           runs have values in a domain-wise table (e.g. ``coverage`` is
           absolute, whereas ``expansions`` is not, because we can't
           compare algorithms A and B for task X if B has no value for
           ``expansions``).
-        * *min_wins*: Set to True if a smaller value for this attribute
+        * *min_wins*: set to True if a smaller value for this attribute
           is better, to False if a higher value is better and to None
           if values can't be compared. (E.g., *min_wins* is False for
           ``coverage``, but it is True for ``expansions``).
-        * *functions*: Set the function or functions used to group
-          values of multiple runs for this attribute. The first entry
-          is used to aggregate values for domain-wise reports (e.g. for
-          ``coverage`` this is :py:func:`sum`, whereas ``expansions``
-          uses :py:func:`geometric_mean`). This can be a single
-          function or a list of functions and defaults to
-          :py:func:`sum`.
-        * *scale*: Default scaling. Can be one of "linear", "log" and
+        * *function*: the function used to aggregate
+          values of multiple runs for this attribute, for example, in
+          domain reports. Defaults to :py:func:`sum`.
+        * *functions*: deprecated. Pass a single *function* instead.
+        * *scale*: default scaling. Can be one of "linear", "log" and
           "symlog". If *scale* is None (default), the reports will
           choose the scaling.
-        * *digits*: Number of digits after the decimal point.
+        * *digits*: number of digits after the decimal point.
 
         The ``downward`` package automatically uses appropriate
         settings for most attributes.
@@ -113,14 +130,15 @@ class Attribute(str):
         """
         self.absolute = absolute
         self.min_wins = min_wins
-        self.functions = tools.make_list(functions)
+        self.function = get_aggregation_function(
+            function, tools.make_list(functions)) or sum
         self.scale = scale
         self.digits = digits
 
     def copy(self, name):
         return Attribute(
             name, absolute=self.absolute, min_wins=self.min_wins,
-            functions=self.functions, scale=self.scale, digits=self.digits)
+            function=self.function, scale=self.scale, digits=self.digits)
 
 
 class Report(object):